pyproject.toml
[project]
name = "spiriter"
version = "0.1.0"
description = "2D sprite background removal and layout tool"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.34.0",
  "python-multipart>=0.0.12",
  "pillow>=11.1.0",
  "numpy>=2.2.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false
